move AssetInfo key implementation
